Temporarily slightly break loading and saving
This is to test the audit tests. Different modes than expected are
passed to open. This change (at least in _load_json) wouldn't be
reasonable to use for real, but it passes (and should pass) all the
other tests.

This commit, whose sole purpose is to serve as a narrow check of
the audit tests for the open event, will be reverted immediately.