Listen for open with less generalization (EliahKagan#11)
I'm not sure if this code is preferable, but it's about half the
size. Because it hard-codes the open event, it also makes sense
with OpenEvent defined in the _audit submodule. It seems to me that
the weaker separation of concerns is inelegant. But this *might* be
regarded better on account of removing more code from test_cached.