Install poetry in onCreate, and include plugins
This is instead of with a dev container feature. In both cases,
pipx is actually used. But for adding plugins, and for making it so
the user of the dev container can easily add more plugins (without
sudo), it's simpler to just do it with commands in onCreate.

I don't think this significantly increases the time it takes to
build the dev container (because each dev container feature
contributes to the total time, too), but if so, that might be a
reason to go back to using the dev container feature (and modifying
the permissions in onCreate, as well as installing the plugins).