Add photo of diagram from design discussion
This shows a couple of ideas about how to reorganize the class
hierarchy in the tests, and how to add classes for more tests, to
test the behaviors currently asserted in test_embed for the
disk-caching versions (currently tested in test_cached) as well,
without duplicating much code or test logic.

This image should probably be removed later, along with notes.md,
but it will be useful for completion of work on the tests branch.