Make audit event handling shorter and nicer
This uses a Mock as a listener and asserts a call of the correct
form.

In an ideal world (for our tests), it would be possible to
subscribe and unsubscribe listeners for specific audit events,
which would be called with the event's args (as separate
arguments). We could just subscribe a unittest.mock.Mock object,
and use its method for asserting that it was called at least once
with the correct values of the arguments we're interested in and
any value of the argument we're not interested in.

This commit changes the _audit submodule so that it allows the test
code to work that way. Extracting to a list is unnecessary, as is
picking out which arguments we want and choosing a data structure
to represent them; the key insight here, which I hadn't had when I
wrote previous versions of our audit event checking code, is that
unittest.mock.Mock objects basically do those two things for us.