Keep specific Mock details out of _audit
The call to the Mock's assertion method, and the use of the ANY
placeholder for Mock calls, are in test_cached_caching, and nothing
else in _audit has to know anything about how the listener is
implemented. So I think it's better to divide it that way.

I've made the context manager pass the listener through in
__enter__, to retain the earlier notational convenience. (This is
inspired by the convenience of how unittest.mock.patch with an
explicit new object still passes that object through.)

This commit also shortens a long variable name.