Use enterContext
Python 3.11 has unittest.TestCase.enterContext. This uses that on
3.11, and defines and uses a simplified enterContext sufficient for
our purposes when the Python version is less than 3.11.

(This also removes some FIXMEs about eliminating tearDown in a
couple leaf classes where there is no disadvantage of using it, and
where the needed logic is not naturally expressed with enterContext
or addCleanup.)