Fix "CI/CD in forks" to not make too-strong claim (#92)
Closes #84

The old wording in the readme wrongly implied that workflows a fork
inherits are *never* enabled automatically or by any mechanism
other than enabling them manually in the Actions tab. That is
inaccurate for reasons detailed in #84 and the sources it
references.

I think this new wording is also more readable than before, even
separately from that issue.