Update "What's here" (#97)
In the readme, this brings the "What's here" section up to date on
the notebooks, and on what I think are the most useful modules to
know about when *starting* to read the code.

In the test suite, some "non-public" modules might be considered
more important than those listed. But I haven't listed them, at
least for now. This is not because there's any reason readers
shouldn't be interested in them. Rather they can be discovered by
following references upward from the listed modules. More
importantly, I don't want this section of the readme to become
excessively long, which would decrease its value and also make it
more prone to falling out of date.