Start on using data frames for cost computation
This uses Polars, but I might switch to pandas, which is already an
indirect dependency of the project, which produces nicer looking
tables by default (albeit in part by omitting explicit type names),
and which is more willing to sum Decimal values, though I worry
this may be doing so by converting to float64 (and thus sometimes
losing precision in a way that is undesirable for monetary values).

The code here is not complete or altogether working, but I'm
committing this before trying out pandas for this, in case I switch
it to pandas but we want to revisit the Polars code later.