Pull dependencies and config out of lint workflow
The pylint job in the lint CI workflow uses three pylint-related
packages that are effectively dev dependencies but that were not
listed in environment.yml or pyproject.toml / poetry.lock. (This is
in addition to pylint itself, which we did list.) It previously
took care of installing and configuring them, and the configuration
modified pyproject.toml on CI, which was complicated and brittle
(and a workaround for the worst of its brittleness added even more
complexity).

This lists them as pip dependencies in environment.yml (conda
packages aren't available for them, so they're in the "pip" group),
and as dev dependencies in pyproject.toml/poetry.lock. One of them
requires configuration, which is now in pyproject.toml. The
workflow file is accordingly much simpler, clearer, and shorter.

The reason for the old way was for version compatibility. When we
first started using pylint-actions, it required Python 3.9+, but
since https://github.com/skhomuti/pylint-actions/releases/tag/0.4.0
it supports all versions of Python that pylint supports, so it can
simply be listed in our manifest files.

In addition to simplifying CI workflow files, this is also much
clearer, allows dependency management tools to detect these
dependencies, and allows these tools to be be tested locally or in
a codespace (which could occasionally be useful when investigating
something related to them, either in them or in our code).