Extract count_tokens to the embed module
This defines count_tokens as a function in embed, includes a TODO
comment noting the problem where it requires lots of memory (which
is why it often failed in a codespace with 4 GiB of RAM, but works
reliably with 8 GiB), and adds tests for it in test_embed.

The notebooks that use count_tokens now use the function defined in
embed and no longer import, nor directly use, tiktoken. (Of course,
the embed module now *does* import and use tiktoken.)

Counting tokens is a generally useful feature, and it seems to me
that it belongs directly in embed for the same reason the
MODEL_DIMENSION constant does.