Make clearer where tracked data/ files come from
This adds headings and a missing example in cached.ipynb, and
divides it into sections with a (very) small amount of explanation.

Previously, only the "hello" embedding saved in data/ was actually
retrieved in cached.ipynb. This adds ["hello", "goodbye"] to the
notebook, which is the other .json file tracked by Git in data/.