Produce JSON without excessive precision
The model produces embeddings with 32-bit coordinates. They are
transmitted without extra precision, and we use dtype=np.float32 to
represent them as NumPy arrays, but when they are represented as
Python lists, the coordinates are Python floats, which are float64.

To use the standard-library json module, we were turning them into
lists of Python floats before converting them to JSON. This caused
their JSON representations in base-10 to have lots of extra digits,
using about twice as much space as necessary.

That we were doing this was related to limitations in the json
module that makes it cumbersome (though not impossible) to output
JSON in which each coordinate has exactly as many digits as
necessary to ensure no loss of precision when round-tripping from
float32.

This commit changes the code in cached.py to output such shorter
JSON. It uses the orjson library, which supports directly
converting NumPy arrays to JSON rather than requiring them to be
made into Python lists first.

This also decreases the indentation ("tab width") in emitted JSON
from 4 to 2. Now that the numbers' representations are so much
shorter, the whitespace accounts for a significant amount of the
size, so I think this is also beneficial.

We may want to go further than that and output JSON without
indentation, with the entire object represented on a single line.
This would be hugely less human-readable, but it may save enough
space to make a difference, in a significant minority of cases, in
whether our JSON output is acceptable way to store many embeddings.

The two .json files in data/ are regenerated, so that the
difference between the old and new JSON representations can be
seen. The .json files in tests_data/ should also be regenerated, so
that our tests more thoroughly test the representation that will be
produced and used. But I have omitted that change from this commit,
to verify and demonstrate (via CI) that the new code is able to
correctly read and use embeddings serialized and saved by the old
code (since, if it couldn't, that would indicate that something
were wrong).