Use orjson in text_or_texts hashing too
We use JSON in two ways (both before and after this change):

1. As the format to save *output*, i.e., embeddings, to disk.

2. As a temporary representation of the *input* whose sole purpose
   is to have its blake3 hash taken to produce theoretically
   non-unique but extremely low-collision cache "keys," from which
   we build filenames.

These are completely separate, and it works to use orjson for (1)
while continuing to use the standard-library json module for (2).
However, orjson offers a benefit for (2) as well: We need JSON as
bytes rather than str to compute its blake3 hash, so we converted
to str. But orjson outputs bytes instead of str. Furthermore, we
are already using it in that module. So using it for (2) slightly
simplifies things.

More important than that, using it for both eliminates the need to
remember which JSON library is being used for which thing, and I
think it makes the code of embed.cached easier to understand.
Furthermore, orjson may be significantly faster (it usually is).

Using json.dumps + encode *sometimes* produces a slightly different
serialization than using orjson.dumps, due to differences in
default style (even when producing unindented JSON). This is why
some of the filenames are different now. References in cached.ipynb
and the automated tests are updated accordingly.