Tell pylint it's OK it can't see into the orjson module
This marks the module's attributes as being dynamically generated.
Except in the sense that all objects and attributes come into
existence at runtime in Python, the situation is kind of the
opposite of dynamic generation: the module is implemented in Rust,
and while I believe it provides a Python fallback, the Rust
implementation is being used, and pylint can't see into that.

There may be a better way to get pylint not to complain about this,
because the module has type stubs (they ship with it in the newest
versions, including the versions used here). I'm not sure.