Unextract _compute_blake3_hash
This removes the _compute_blake3_hash helper function, replacing it
with a slightly less confusing way to doing the pylint suppression
and having it apply to just one statement (still without exceeding
our preferred line length limitations).

The issue with _compute_blake3_hash is that it seemed trivial, and
the reason for its existence was not necessarily obvious even by
carefully reading. This could be addressed by expanding its
docstring or adding another comment, but I think this change is
slightly better than doing that.