Go back to using "x" instead of "w"
If the file already existed but couldn't be read, then we should
fail either with that error or an error that will make clear that
it occurred. Passing "x" in mode does the latter. Passing "w" in
mode makes the code simpler, but does neither of those, so I don't
think it's really suitable.