Only catch FileNotFoundError
Other causes/subclasses of OSError mostly cannot be addressed by
attempting to create the file. This causes them to fail earlier,
producing less complicated tracebacks for them and avoiding extra
work.

It also allows them to be separately handled by an application,
which could catch other subclasses of OSError or just catch OSError
itself to handle conditions where attempting to open the file fails
for reasons other than it not being found.

This also more clearly expresses the condition we are branching on
in the except vs. else clause, so it may improve code readability.