Simplify writing file, now that it's safe enough
When OSError was caught, the likelihood of attempting to write a
file when something already existed was higher.

Even though we catch FileNotFoundError specifically now, rather
than the more general OSError exception, there could potentially be
a benefit to opening the file for create only. But the benefit is
less, and the code is shorter and simpler with 'w' instead of 'x'.
This is because Path has a write_bytes method corresponding to its
read_bytes method already being used. This uses that method, and
changes the tests to expect 'w' rather than 'x'. (We might
ultimately want the tests to allow either one.)