Copy .safetensors files from tests_data if present
Currently no .safetensors files are saved in the tests_data
directory, so this makes no difference, but when they are added,
they will be copied when tests in test_cached_embeddings run.

This is far from the only change that needs to be made in
test_cached_embeddings to test safetensors as well as JSON. But I
think I'll expand the tests in test_cached_caching to test both
JSON and safetensors before expanding test_cached_embeddings, so
I'm committing this change by itself.