Start on test_cached_embeddings safetensors tests
They appear to work, but they actually do *not* work correctly,
because both the JSON and safetensors tests are using JSON. This is
because the file_type property is never actually being used in
these tests.

I plan to change the design in embed.cached so that the default
encoding is read from a module attribute. Such a design is more
readily testable because it is easy to monkey-patch such an
attribute in the same way the test_cached_embeddings tests are
already monkey-patching DEFAULT_DATA_DIR.