Update binary file patterns in .gitattributes
This makes Git treat files with a .safetensors suffix as binary
even if it would otherwise auto-detect them as text (since they are
never text). It is the same precaution used for .pdf and previously
used for .jpg.

This change ought to have been done in 4d669a0 when .safetensors
files were first committed to the repository, but inspecting output
of "git ls-files --eol" (https://stackoverflow.com/a/66796286)
shows that none have been misdetected as text so far.

This also removes the listing for .jpg because we no longer have
any .jpg files in the project (and we're not trying to exhaustively
list all possible binary file suffixes).