Add notebook diff tools (#145)
This installs nbdime as a dependency in the notebook group. The
most useful tool is nbdiff-web. When VS Code is unavailable or
inconvenient to use, this will show diffs. Now that we've added
jupyterlab and related packages to the notebook group, I think it
makes sense to have this too.