Fix JupyterLab CORS in codespaces
This matches on the domain name generated from the specific
codespace name, which is determined in a postStart script.

The script is copied from palgoviz. Currently it is the code
shown here:

https://github.com/EliahKagan/palgoviz/blob/main/.devcontainer/postStart

I believe what I wrote that will work just as well here.

This doesn't fix the similar but separate CORS problem with
nbdiff-web in codespaces. (While they are often used together,
jupyterlab and nbdime are separate.)