Make regex matching logic more idiomatic
Although the previous way has the advantage of making clear that
the regular expressions are used for full matches, it is somewhat
unidiomatic. I like that style, but I think assigning the results
of re.compile to constants and calling fullmatch at the use sites
will be easier to understand, for people unfamiliar with the code,
than "precomputing" their bound fullmatch methods and assigning
those as helpers.