Make TestDiskCachedBase._dir_path public (#154)
This renames it to be public, from _dir_path to dir_path.

It is accessed, and needs to be accessed, in derived classes. So it
should be public or protected. I don't think the additional
complexity of it being protected, even just documenting it as such,
is justified here. Making it public also is more consistent with
other design choices where members the test bases provide to their
derived classes are public in our test suite.

See #153, which this fixes, for details.

(Note that this is still a member of a class that is contained in a
nonpublic module, _bases.)