Put more metadata in pyproject.toml (#162)
For when we eventually publish to PyPI.

Of course, since we are going to rename the project sometime
between now and when we publish to PyPI, we'll have to change the
metadata that reflect the name. But I think it's still easier to
have most or all of the fields beforehand; the name can be changed
with a supervised search and replace when it's time to do so.