Keep flake8 from scanning where it shouldn't (#166)
This extends flake8's exclusions, adding directories where virtual
environments may be kept (in the less common but still significant
case where it is kept inside the project directory, see #165 for
details), as well as .ipynb_checkpoints (which .py files could
potentially end up in, in some development workflows).

The main goal of this change is to fix #165. In that bug report, I
argued that, for virtual environments, it would be sufficient to
exclude just ".venv". However, in practice, it's almost as easy to
exclude all directories whose names unambiguously indicate that
they are meant to hold a virtual environment, so I did that here.

flake8 does not itself look in pyproject.toml for configuration.
Rather than make another configuration file just for flake8, I've
added the flake8-pyproject plugin, which gives flake8 this feature.