Give explicit Python versions for flake8-pyproject (#169)
Specifying the Python versions for flake8-pyproject as the same
range as for flake8 itself (which is just the range of versions
flake8 actually supports) fixes #168. Without this, calculating
dependencies usually works, but calculating or expressing them
fails at least in the situation of "poetry export --only-dev"
(which should be able to work, CI might use it in the future, and
our tox configuration uses it currently).