Add pytest runner and use it via tox
This adds pytest as a developement dependency, but only for the
purpose of using it as a test runner for our unittest tests. All
the automated tests in the project remain unittest and there are
currently no plans to change that, but it is sometimes convenient
to use pytest as the test runner, because of the way in which it
shows output.

One situation where this is likely to benefit is when running
tests with tox, and this also changes the tox configuration to run
the unittest tests with pytest instead of the unittest runner
itself.

No CI jobs use pytest, and there is currently no plan to change
that.