Keep numpy at 1.24.* for Python 3.8 support
Because numpy requires 3.9 starting in 1.25.0.

Unlike Python 3.7, which is hard to support for multiple other
reasons, and which the PSF will stop maintaining/supporting soon,
Python 3.8 will remain relevant for a while and is, so far, fairly
easy to continue supporting.

This commit also bumps the indirect dependency websocket-client
from 1.5.3 to 1.6.0.