Use latest ipython compatible with Python version
This is analogous to #174, but for ipython. This is less important
than being able to use the latest numpy, but still useful. In
particular, being able to use the newest ipython restores the
right arrow for completing a whole line, which had become broken.

Unlike numpy, ipython was not previously listed in pyproject.toml,
so this is adding that item as a list, rather than changing it to
be a list.