Rename "use_box" parameter to "enable"
This is less confusing, because "use_box" sounds like it would
control whether height bounding would be done with or without a
box, while actually this parameter controls whether height bounding
is done *at all* (and if it's not, no box is used either).

I had originally avoided "enable" because the similar name
"enabled" is used for some widgets with a different meaning, but I
think that possible source of confusion is less severe and also
less likely to occur in the first place.