Our element-to-string conversions include tail text
This adds code to usc_manual.ipynb to demonstrate that the way we
are converting XML nodes (elements) to strings includes any tail
text at the end of the string, even though an element's tail text
conceptually belongs to the element's parent.

This may be good or bad, depending on where we want text after a
subelement that is not itself part of another subelement to go, in
pieces of text whose embeddings we compute. (I suspect that this
behavior of lxml.etree.tostring may help more than hinder.)

In addition to those additions, this reorders the recently added
material in usc_manual.ipynb so that it makes more sense when
reading the notebook top-down (and so the part we are likely to
extend further in the very near future is at the end). This also
refactors some of the code there for brevity and readability.