Let pylint see inside blake3 and orjson (#183)
This is a better and more fine-grained solution to the orjson
linting problem that I'd handled before by marking orjson as
dynamically generated in eab6c18 (it is not dynamically generated).

Similarly, we were suppressing a "not callable" warning on
blake3.blake3, which was due to pylint not importing and looking
inside the blake3 module. This is better approach to that, too.

See:

- https://pylint.pycqa.org/en/v2.17.4/user_guide/messages/error/no-member.html
- https://pylint.pycqa.org/en/v2.17.4/user_guide/messages/error/not-callable.html

Although only the `no-member` documentation (and not the
`not-callable` documentation) explicitly recommends the use of
`extension-pkg-allow-list`, I believe it is the ideal approach to
this particular `not-callable` error as well.

This also bumps direct and indirect dependencies.