Let pylint see inside lxml
This doesn't typically appear needed when pylint is run locally,
but CI produces c-extension-no-member / I1101 for each attribute
access expression on etree (that is, importing etree itself is no
problem, but accessing fromstring, tostring, and iterwalk on etree
produce I1101 diagnostics).

This lists the lxml package for extension-pkg-allow-list, which
already existed to list blake3 and orjson; see 95fa82c (#183).