Handle loose USC archives (as well as nested) (#188)
The official USC archive contains "loose" top-level files rather
than a top-level directory. As the FIXME this takes care of noted,
we didn't support that (expecting a top-level directory). Both work
now.

I've tested this with the repacked (higher compression) archive
we've been using that has a top-level directory ("nested"), as well
as with the official archive that does not ("loose").