Report lost "tail" text in get_embeddable_elements
+ Improve reporting style for "text" text.

It appears no "text" or "tail" text appears in any element
get_embeddable_elements would split up, in Title 42.

We should probably provide a keyword argument to make it an error
if it does happen, though, if we're not going to add logic to
preserve such text in a situation where we would break up an
element that has it. Such a "strict" mode should also apply to a
too-big leaf, another condition we don't currently encounter but
that would be skipped.

This commit does *not* itself add any such strict mode. For this
reason, one of the related FIXMEs is retained.