Have get_embeddable_elements raise on lost content (#190)
This modifies get_embeddable_elements to add a strict mode, which
is enabled by default and which, after computing the entire
breakdown, checks if *any* of the three situations in which content
would be lost has occurred. If so, it raises ValueError, reporting
how many times each of the three general situations occurred. (The
idea behind it being ValueError is that all these situations occur,
or not, based on the value of the input, not its type, or the
result of attempting to look something up, etc.)

This also upgrades the lost "text" text and lost "tail" text
situations from warning to error in the logging done immediately
when the problems are discovered, so all three have "error"
severity.