Do extra validation as zipfile module docs suggest
This checks the namelist for paths that look like they may be
usable to perform a directory traversal attack. It is not likely to
be more robust than the checks zipfile.ZipFile.extractall is
already equipped with, but it treats these as *errors* and refuses
to use the archive, rather than using the archive but modifying the
paths accordingly. It does this without excessive complexity by
relying, to some extent, on knowledge about what can reasonably
go in a USC archive (that is, it can afford to reject some
archives that would be reasonable in other contexts).

It also covers the unintuitive issue of how some paths on Windows
are not regarded to be absolute (at least by pathlib.Path) even
though they can access outside the current directory even without
anything like ".." ever appearing as a path component. I *think*
the ZipFile class's extract and extractall methods do cover this,
but I'm not sure whether it strictly follows from the documentation
that all conforming Python implementations would necessarily do so.

Another reason for this change is to be in line with the strong
encouragement in the documentation to inspect archives before
extracting them. The stakes here are low, because USC archives are
unlikely ever to be totally untrusted, but I think the increase in
code length and complexity is not excessive.