Put some iterwalk-based code in the usc module
This extracts repeated ET.iterwalk logic from usc_42.ipynb to
usc.py, as the new walk_tag function, modified to find the schema
prefix from the root (rather than relying on having it as a global,
as was done in the notebook).

This also extracts get_direct_sections, having it use walk_tag to
get its iterwalk iterator.