Remove notes_tag, which is no longer needed
This deletes the cell in usc_42.ipynb that introduced notes_tag,
because the code that used it instead uses facilities in usc.py
that compute it as needed.

In contrast, there is a small amount of use of section_tag in
usc_42.ipynb, so schema_prefix and section_tag remain there.