Revert "Remove notes_tag, which is no longer needed"
One of the changes I made in #198, removing the notes_tag global
variable from usc_42.ipynb, was incorrect, because there was still
one place in that notebook that used it. I'm not sure how I missed
that; maybe I failed to rerun the notebook after that change, since
the following change was outside the notebook. In any case, while
it will eventually be removable from the notebook, that was not
(and is not) yet the case. This puts it back so the notebook works
again.

This reverts commit 2cf0a03662052e2c75f0b2741809cd63b5b0a5a3.