Re-run usc_42.ipynb with notes_tag restored
Ordinarily I would have included this in the commit it tests,
rather than having it as a separate commit. In this case, though,
I want to make it especially easy to check the diffs, to avoid a
repeat of the same kind of mistake I made in prematurel removing
the notes_tag global variable before.