Have USC notebooks use data/usc/ instead of data/
Whatever automated tests we make of the USC demo should probably do
the same. Note that this does not change any defaults for what code
in the embed.demos.usc module does if data_dir is not passed. It
only modifies the callers (currently all in notebooks) to specify
the usc subdirectory of the repo's data directory.