Rename CONTEXT_LENGTH to TOKEN_LIMIT (#204)
* Rename CONTEXT_LENGTH to TOKEN_LIMIT

In a completion model, the total number of tokens the model can
work with is nearly always called its context length. But in an
embedding model, the terminology is less uniform. Since the phrase
"context length" may be unclear to some developers, and furthermore
since it is less intuitive in the setting of text embeddings where
it is less often meaningful to regard the input (or some part of
it) as "context," I've renamed embed.CONTEXT_LENGTH to
embed.TOKEN_LIMIT.

All uses in the project are updated. It's possible that a third,
even clearer name will be found in the future; I do not claim that
TOKEN_LIMIT is necessarily the best name for this.

This also expands the TOKEN_LIMIT docstring to explain specifically
how it applies to calls to both kinds of our embedding functions.

* Re-run usc_42.ipynb to test rename

Unlike in #203, where the second commit should possibly be included
in the target branch when merging, in this case this commit can
definitely be squashed into the one before it with no disadvantage.