Update scipy
Before this, scipy 1.9.3 was being installed. This is due to the
upper bound more recent versions of scipy specify:

- scipy 1.9.3 requires Python >=3.8
- scipy 1.10.1 requires Python <3.12,>=3.8
- scipy 1.11.1 requires Python <3.13,>=3.9

I don't think older versions of SciPy really support later (and
forthcoming) versions of Python, but instead that it was only in
1.10.* that SciPy began to explicitly state a required upper bound
on the Python version. But that was causing poetry to select scipy
1.9.3 as the best version to satisfy our dependencies, so that
scipy version (and no others) was listed in poetry.lock. This
happened because our project's upper bound was higher.

This doesn't happen with conda (at least not in this project). Our
environment.yml file specifies python=3.11, and scipy 1.11.1 is
selected. When a lower minor version of Python is specified in
environment.yml, as we do on CI (by editing environment.yml
automatically and not committing the change), it still works fine:
versions of Python that work with SciPy 1.11.1 select that, and
versions that don't work with 1.11.* but do work with 1.10.1 select
that. Therefore, it is not necessary to add scipy to
environment.yml to get a current version when using conda/mamba.

The scipy dependency of our project is indirect, through
openai[embeddings], because openai.embeddings_utils uses it in two
ways:

- Directly, in one place, for its implementations of a few
  different distance functions of interest.

- Indirectly, because scikit-learn declares scipy as a dependency,
  and embeddings_utils uses scikit-learn in several places for
  various functionality.

We are not currently making use of functionality provided (even
indirectly) by scipy or scikit-learn. However:

- Once we rename this project and publish a wheel on PyPI, it may
  be used together with such functionality.

- We are moderately likely to use some functionality in
  embeddings_utils that makes use of scikit-learn, in the future,
  in one or more of the demos ("small app examples").

There are disavantages of changing our supported Python version
from ^3.8 (i.e., >=3.8,<4.0) to >=3.8,<3.13, as I do here:

- It has to be changed again to support Python 3.13 and later when
  they come out.

- My preliminary experiments suggest that we cannot yet actually
  support even Python 3.12 (which is in beta at this time). We'll
  eventually be able to support it, probably with no effort on our
  part (the issue is some of our other dependencies). But giving
  the upper bound as <3.13 could mislead users into thinking this
  software should be expected to work on some existing 3.12 beta.

However, specifying a lower upper bound to address the second of
those disadvantages would add further complexity, and also
exacerbate the first of those disadvantages.

Given that it seems there is no perfect solution, and that we have
not yet packaged this for PyPI and we are not yet making use of
scipy or scikit-learn, it may seem that the best approach is to let
scipy 1.9.3 continue to be selected, for now. However, that carries
its own disadvantage: we would have to handle CVE-2023-25399 in a
different way from the usual approach of patching as soon as
patching is feasible.

Deferring action on CVE-2023-25399 might be defensible from a
security perspective, since it is unclear that it really
constitutes a security bug and it seems the CVE may be been
requested in error. On that, see
https://github.com/scipy/scipy/issues/16235#issuecomment-1625361328
and its replies, which appear to be in agreement. However, it is
still simpler to patch (and to do so in way that 1.9.3 no longer
appears as any alternative in poetry.lock) than defer action on it.

(At this time, I don't feel comfortable recommending that we mark
it as decisively not affecting us. The related CVE-2023-29824 is
noted as disputed, saying "the vendor and discoverer indicate that
this is not a security issue". In contrast, CVE-2023-25399 does not
currently contain any such text.)

Due to a combination of the above factors, I've added scipy as a
direct dependency in pyproject.toml so that current versions are
selected during dependency resolution. We might revisit this later.