Install pylint-sarif-unofficial via pipx (#209)
Without this change, pylint-sarif-unofficial should still start
working again, because version 0.2.0 holds back the version of
jsonschema. However, we have other dependencies that pull in
jsonschema as an indirect dependency, and I'd rather not keep those
from benefiting from improvements.

* Install pylint-sarif-unofficial via pipx

And use the new version.

This change--both installing it with pipx on CI instead of having
it as a development dependency, and using the new version--is to
work around https://github.com/EliahKagan/pylint-sarif/issues/19.

This also upgrades (other) direct and indirect dependencies.

* Use stable 0.2.0

(This version is actually the same code as the alpha.)

* Don't specify version

0.2.0 is available when given explicitly, as the previous commit
verifies. It should be selected automatically when no version is
given, as long as no later stable version is available.