Reset CodeQL advanced config to current default
This updates it to what is currently automatically generated for a
new configuration. It also removes the setup-python step, which we
added because we have been having it install Python dependencies
and wanted the latest stable Python we support to be used for that.

GitHub now regards CodeQL to work well even without installing
Python dependencies, and recommends turning that off. This commit
is the first step to making that change in a way that makes sense
for this project.